requirements.txt
fastapi>=0.100.0
strawberry-graphql[fastapi]>=0.256.0,<0.260.0
uvicorn>=0.20.0
pydantic>=2.7.0,<2.10.0
